+++ R/ggbarstats.R
@@ -1,53 +1,33 @@
-#' @title Bar(column) charts with statistical tests
+#' @title Bar (column) charts with statistical tests
 #' @name ggbarstats
 #' @aliases ggbarstats
 #' @description Bar charts for categorical data with statistical details
 #'   included in the plot as a subtitle.
-#' @author Chuck Powell
+#' @author Chuck Powell, Indrajeet Patil
 #'
-#' @param factor.levels A character vector with labels for factor levels of
-#'   `main` variable.
-#' @param title The text for the plot title.
-#' @param caption The text for the plot caption.
+#' @param labels.legend A character vector with custom labels for levels of
+#'   the `main` variable displayed in the legend.
+#' @param xlab Custom text for the `x` axis label (Default: `NULL`, which
+#'   will cause the `x` axis label to be the `main` variable).
+#' @param ylab Custom text for the `y` axis label (Default: `"percent"`).
+#' @param bar.proptest Decides whether proportion test for `main` variable is
+#'   to be carried out for each level of `condition` (Default: `TRUE`).
+#' @param data.label Character decides what information needs to be displayed
+#'   on the label in each pie slice. Possible options are `"percentage"`
+#'   (default), `"counts"`, `"both"`.
 #' @param legend.position The position of the legend
-#'   "none", "left", "right", "bottom", "top" (Default: `"bottom"`).
-#' @param x.axis.orientation The orientation of the x axis labels one of
+#'   `"none"`, `"left"`, `"right"`, `"bottom"`, `"top"` (Default: `"bottom"`).
+#' @param x.axis.orientation The orientation of the `x` axis labels one of
 #'   "slant" or "vertical" to change from the default horizontal
 #'   orientation (Default: `NULL` which is horizontal).
-#' @param sample.size.label Logical that decides whether sample size information
-#'   should be displayed for each level of the grouping variable `condition`
-#'   (Default: `TRUE`).
-#' @param palette If a character string (e.g., `"Set1"`), will use that named
-#'   palette. If a number, will index into the list of palettes of appropriate
-#'   type. Default palette is `"Dark2"`.
-#' @param facet.wrap.name The text for the facet_wrap variable label.
-#' @param facet.proptest Decides whether proportion test for `main` variable is
-#'   to be carried out for each level of `condition` (Default: `FALSE`).
-#' @param perc.k Numeric that decides number of decimal places for percentage
-#'   labels (Default: `0`).
 #' @param label.text.size Numeric that decides size for bar labels
 #'   (Default: `4`).
 #' @param label.fill.color Character that specifies fill color for bar labels
 #'   (Default: `white`).
 #' @param label.fill.alpha Numeric that specifies fill color transparency or
-#'   `"alpha"` for bar labels (Default: `1` range 0 to 1).
-#' @param slice.label Character decides what information needs to be displayed
-#'   on the label in each sub-divsion of the vertical bar. Possible options are `"percentage"`
-#'   (default), `"counts"`, `"both"`.
-#' @param bf.message Logical that decides whether to display a caption with
-#'   results from bayes factor test in favor of the null hypothesis (default:
-#'   `FALSE`).
-#' @param ggplot.component A `ggplot` component to be added to the plot prepared
-#'   by `ggstatsplot`. This argument is primarily helpful for `grouped_` variant
-#'   of the current function. Default is `NULL`. The argument should be entered
-#'   as a function. If the given function has an argument `axes.range.restrict`
-#'   and if it has been set to `TRUE`, the added ggplot component *might* not
-#'   work as expected.
-#' @inheritParams bf_contingency_tab
-#' @inheritParams subtitle_contingency_tab
-#' @inheritParams subtitle_onesample_proptest
-#' @inheritParams paletteer::scale_fill_paletteer_d
-#' @inheritParams theme_ggstatsplot
+#'   `"alpha"` for bar labels (Default: `1` range `0` to `1`).
+#' @param bar.outline.color Character specifying color for bars (default: `"black"`).
+#' @inheritParams ggpiestats
 #'
 #' @import ggplot2
 #'
@@ -59,18 +39,9 @@
 #' @importFrom groupedstats grouped_proptest
 #' @importFrom tidyr uncount
 #' @importFrom tibble as_tibble
+#' @importFrom scales percent
 #'
-#' @references
-#' \url{https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/ggbarstats.html}
-#'
-#' @return Unlike a number of statistical softwares, `ggstatsplot` doesn't
-#'   provide the option for Yates' correction for the Pearson's chi-squared
-#'   statistic. This is due to compelling amount of Monte-Carlo simulation
-#'   research which suggests that the Yates' correction is overly conservative,
-#'   even in small sample sizes. As such it is recommended that it should not
-#'   ever be applied in practice (Camilli & Hopkins, 1978, 1979; Feinberg, 1980;
-#'   Larntz, 1978; Thompson, 1988). For more, see-
-#'   \url{http://www.how2stats.net/2011/09/yates-correction.html}
+#' @inherit ggpiestats return details
 #'
 #' @examples
 #'
@@ -84,7 +55,7 @@
 #'   condition = cyl,
 #'   bf.message = TRUE,
 #'   nboot = 10,
-#'   factor.levels = c("0 = V-shaped", "1 = straight"),
+#'   labels.legend = c("0 = V-shaped", "1 = straight"),
 #'   legend.title = "Engine"
 #' )
 #'
@@ -104,14 +75,15 @@ ggbarstats <- function(data,
                        main,
                        condition = NULL,
                        counts = NULL,
-                       ratio = NULL, #does nothing
+                       ratio = NULL,
                        paired = FALSE,
-                       factor.levels = NULL,
+                       labels.legend = NULL,
                        stat.title = NULL,
                        sample.size.label = TRUE,
                        label.text.size = 4,
                        label.fill.color = "white",
                        label.fill.alpha = 1,
+                       bar.outline.color = "black",
                        bf.message = FALSE,
                        sampling.plan = "jointMulti",
                        fixed.margin = "rows",
@@ -125,11 +97,12 @@ ggbarstats <- function(data,
                        simulate.p.value = FALSE,
                        B = 2000,
                        legend.title = NULL,
-                       facet.wrap.name = NULL,
+                       xlab = NULL,
+                       ylab = "Percent",
                        k = 2,
                        perc.k = 0,
-                       slice.label = "percentage",
-                       facet.proptest = FALSE,
+                       data.label = "percentage",
+                       bar.proptest = TRUE,
                        ggtheme = ggplot2::theme_bw(),
                        ggstatsplot.layer = TRUE,
                        package = "RColorBrewer",
@@ -148,12 +121,12 @@ ggbarstats <- function(data,
   if (is.null(legend.title)) {
     legend.title <- rlang::as_name(rlang::ensym(main))
   }
-  # if facetting variable name is not specified, use the variable name for
-  # 'condition' argument
-  if (is.null(facet.wrap.name)) {
-    facet.wrap.name <- rlang::as_name(rlang::ensym(condition))
-  }
 
+  # if alternate variable label is not specified, use the variable name for
+  # 'condition' argument
+  if (is.null(xlab)) {
+    xlab <- rlang::as_name(rlang::ensym(condition))
+  }
 
   # =============================== dataframe ================================
 
@@ -186,7 +159,6 @@ ggbarstats <- function(data,
 
   # main and condition need to be a factor for this analysis
   # also drop the unused levels of the factors
-
   data %<>%
     dplyr::mutate(.data = ., main = droplevels(as.factor(main))) %>%
     dplyr::filter(.data = ., !is.na(main)) %>%
@@ -197,62 +169,60 @@ ggbarstats <- function(data,
   data %<>%
     tibble::as_tibble(x = .)
 
+  # convert the data into percentages; group by conditional variable
   df <-
     data %>%
     dplyr::group_by(.data = ., condition, main) %>%
     dplyr::summarize(.data = ., counts = n()) %>%
     dplyr::mutate(.data = ., perc = (counts / sum(counts)) * 100) %>%
     dplyr::ungroup(x = .) %>%
-    dplyr::arrange(.data = ., dplyr::desc(x = main))
-
+    dplyr::arrange(.data = ., dplyr::desc(x = main)) %>%
+    dplyr::filter(.data = ., counts != 0L)
 
   # checking what needs to be displayed on pie slices as labels also decide on
   # the text size for the label; if both counts and percentages are going to
   # be displayed, then use a bit smaller text size
-  if (slice.label == "percentage") {
+  if (data.label == "percentage") {
     # only percentage
     df %<>%
       dplyr::mutate(
         .data = .,
-        slice.label = paste0(round(x = perc, digits = perc.k), "%")
+        data.label = paste0(round(x = perc, digits = perc.k), "%")
       )
-
-  } else if (slice.label == "counts") {
+  } else if (data.label == "counts") {
     # only raw counts
     df %<>%
       dplyr::mutate(
         .data = .,
-        slice.label = paste0("n = ", counts)
+        data.label = paste0("n = ", counts)
       )
-
-  } else if (slice.label == "both") {
+  } else if (data.label == "both") {
     # both raw counts and percentages
     df %<>%
       dplyr::mutate(
         .data = .,
-        slice.label = paste0(
+        data.label = paste0(
+          "n = ",
+          counts,
+          "\n(",
           round(x = perc, digits = perc.k),
-          "% (",
-          counts, ")"
+          "%)"
         )
       )
   }
 
-#  return(df)
-
   # ============================ sample size label ==========================
 
   # if labels are to be displayed on the bottom of the charts
-  # for each facet
+  # for each bar/column
   if (isTRUE(sample.size.label)) {
     df_n_label <- data %>%
-      group_by(condition) %>%
-      summarize(N = n()) %>% mutate(N = paste0("n = ",N))
+      dplyr::group_by(.data = ., condition) %>%
+      dplyr::summarize(.data = ., N = n()) %>%
+      dplyr::mutate(.data = ., N = paste0("(n = ", N, ")", sep = ""))
   }
 
-#  return(df_n_label)
-
-  # ================= preparing names for legend and facet_wrap ==============
+  # ====================== preparing names for legend  ======================
 
   # reorder the category factor levels to order the legend
   df$main <- factor(
@@ -261,23 +231,14 @@ ggbarstats <- function(data,
   )
 
   # getting labels for all levels of the 'main' variable factor
-  if (is.null(factor.levels)) {
+  if (is.null(labels.legend)) {
     legend.labels <- as.character(df$main)
-  } else if (!missing(factor.levels)) {
-    legend.labels <- factor.levels
-  }
-
-  # custom labeller function to use if the user wants a different name for
-  # facet_wrap variable
-  label_facet <- function(original_var, custom_name) {
-    lev <- levels(x = as.factor(original_var))
-    lab <- paste0(custom_name, ": ", lev)
-    names(lab) <- lev
-    return(lab)
+  } else if (!missing(labels.legend)) {
+    legend.labels <- labels.legend
   }
 
   # =================================== plot =================================
-#  return(df)
+
   # if no. of factor levels is greater than the default palette color count
   palette_message(
     package = package,
@@ -285,28 +246,42 @@ ggbarstats <- function(data,
     min_length = length(unique(levels(data$main)))[[1]]
   )
 
-  p <- ggplot(data = df,
-              aes(fill=main, y=perc, x=condition)) +
-              geom_bar(stat="identity", position="fill", color = "gray") +
-              ylab("Percent") +
-              xlab(facet.wrap.name) +
-              scale_y_continuous(labels = scales::percent, breaks = seq(0, 1, by = 0.10)) +
-              geom_label(aes(label = slice.label, group = main),
-                        show.legend = FALSE,
-                        position = position_fill(vjust = 0.5),
-                        size = label.text.size,
-                        fill = label.fill.color,
-                        alpha = label.fill.alpha
-                        ) +
-              ggstatsplot::theme_mprl(
-                        ggtheme = ggtheme,
-                        ggstatsplot.layer = ggstatsplot.layer
-               ) +
-              theme(panel.grid.major.x = element_blank(),
-                    legend.position = legend.position) +
-              ggplot2::guides(fill = ggplot2::guide_legend(title = legend.title))
+  # plot
+  p <- ggplot2::ggplot(
+    data = df,
+    mapping = ggplot2::aes(fill = main, y = perc, x = condition)
+  ) +
+    ggplot2::geom_bar(
+      stat = "identity",
+      position = "fill",
+      color = bar.outline.color,
+      na.rm = TRUE
+    ) +
+    ggplot2::scale_y_continuous(
+      labels = scales::percent,
+      breaks = seq(from = 0, to = 1, by = 0.10),
+      minor_breaks = seq(from = 0.05, to = 0.95, by = 0.10)
+    ) +
+    ggplot2::geom_label(
+      mapping = ggplot2::aes(label = data.label, group = main),
+      show.legend = FALSE,
+      position = ggplot2::position_fill(vjust = 0.5),
+      size = label.text.size,
+      fill = label.fill.color,
+      alpha = label.fill.alpha,
+      na.rm = TRUE
+    ) +
+    ggstatsplot::theme_mprl(
+      ggtheme = ggtheme,
+      ggstatsplot.layer = ggstatsplot.layer
+    ) +
+    ggplot2::theme(
+      panel.grid.major.x = ggplot2::element_blank(),
+      legend.position = legend.position
+    ) +
+    ggplot2::guides(fill = ggplot2::guide_legend(title = legend.title))
 
-
+  # color palette
   p <- p +
     paletteer::scale_fill_paletteer_d(
       package = !!package,
@@ -314,145 +289,177 @@ ggbarstats <- function(data,
       direction = direction,
       name = "",
       labels = unique(legend.labels)
-      )
-
-
-#  return(p)
-#  return(df)
-
+    )
 
   # =============== chi-square test (either Pearson or McNemar) =============
 
-  # if facetting by condition is happening
+  # if testing by condition is happening
+  if (isTRUE(bar.proptest)) {
+    # merging dataframe containing results from the proportion test with
+    # counts and percentage dataframe
+    df2 <-
+      dplyr::full_join(
+        x = df,
+        # running grouped proportion test with helper functions
+        y = groupedstats::grouped_proptest(
+          data = data,
+          grouping.vars = condition,
+          measure = main
+        ),
+        by = "condition"
+      ) %>%
+      dplyr::mutate(
+        .data = .,
+        significance = dplyr::if_else(
+          condition = duplicated(condition),
+          true = NA_character_,
+          false = significance
+        )
+      ) %>%
+      dplyr::filter(.data = ., !is.na(significance))
 
-    if (isTRUE(facet.proptest)) {
-      # merging dataframe containing results from the proportion test with
-      # counts and percentage dataframe
-      df2 <-
-        dplyr::full_join(
-          x = df,
-          # running grouped proportion test with helper functions
-          y = groupedstats::grouped_proptest(
-            data = data,
-            grouping.vars = condition,
-            measure = main
-          ),
-          by = "condition"
-        ) %>%
-        dplyr::mutate(
-          .data = .,
-          significance = dplyr::if_else(
-            condition = duplicated(condition),
-            true = NA_character_,
-            false = significance
-          )
-        ) %>%
-        dplyr::filter(.data = ., !is.na(significance))
+    # display grouped proportion test results
+    if (isTRUE(messages)) {
+      # capturing names of variables
+      main.name <- rlang::as_name(rlang::ensym(main))
+      condition.name <- rlang::as_name(rlang::ensym(condition))
 
-      # display grouped proportion test results
-      if (isTRUE(messages)) {
-        # tell the user what these results are
-        base::message(cat(
-          crayon::green("Note: "),
-          crayon::blue("Results from faceted one-sample proportion tests:\n"),
-          sep = ""
-        ))
+      # tell the user what these results are
+      base::message(cat(
+        crayon::green("Note: "),
+        crayon::blue("Results from one-sample proportion tests for each\n"),
+        crayon::blue("      level of the variable "),
+        crayon::yellow(condition.name),
+        crayon::blue(" testing for equal\n"),
+        crayon::blue("      proportions of the variable "),
+        crayon::yellow(main.name),
+        crayon::blue(".\n"),
+        sep = ""
+      ))
 
-        # print the tibble and leave out unnecessary columns
-        print(tibble::as_tibble(df2) %>%
-          dplyr::select(.data = ., -c(main:slice.label)))
-      }
+      # print the tibble and leave out unnecessary columns
+      print(tibble::as_tibble(df2) %>%
+        dplyr::select(.data = ., -c(main:data.label)))
     }
+  }
 
-#  return(p)
+  # running approprate statistical test
+  # unpaired: Pearson's Chi-square test of independence
+  subtitle <-
+    subtitle_contingency_tab(
+      data = data,
+      main = main,
+      condition = condition,
+      nboot = nboot,
+      paired = paired,
+      stat.title = stat.title,
+      conf.level = conf.level,
+      conf.type = "norm",
+      simulate.p.value = simulate.p.value,
+      B = B,
+      messages = messages,
+      k = k
+    )
 
-    # running approprate statistical test
-    # unpaired: Pearson's Chi-square test of independence
-    subtitle <-
-      subtitle_contingency_tab(
+  # preparing the BF message for null hypothesis support
+  if (isTRUE(bf.message)) {
+    bf.caption.text <-
+      bf_contingency_tab(
         data = data,
         main = main,
         condition = condition,
-        nboot = nboot,
-        paired = paired,
-        stat.title = stat.title,
-        conf.level = conf.level,
-        conf.type = "norm",
-        simulate.p.value = simulate.p.value,
-        B = B,
-        messages = messages,
+        sampling.plan = sampling.plan,
+        fixed.margin = fixed.margin,
+        prior.concentration = prior.concentration,
+        caption = caption,
+        output = "caption",
         k = k
       )
+  }
 
-    # preparing the BF message for null hypothesis support
-    if (isTRUE(bf.message)) {
-      bf.caption.text <-
-        bf_contingency_tab(
-          data = data,
-          main = main,
-          condition = condition,
-          sampling.plan = sampling.plan,
-          fixed.margin = fixed.margin,
-          prior.concentration = prior.concentration,
-          caption = caption,
-          output = "caption",
-          k = k
-        )
-    }
+  # if bayes factor message needs to be displayed
+  if (isTRUE(bf.message)) {
+    caption <- bf.caption.text
+  }
 
-    # if bayes factor message needs to be displayed
-    if (isTRUE(bf.message)) {
-      caption <- bf.caption.text
-    }
+  # ====================== proportion test =======================
 
-    # ====================== facetted proportion test =======================
+  # adding significance labels to bars for proportion tests
+  if (isTRUE(bar.proptest)) {
+    p <-
+      p +
+      ggplot2::geom_text(
+        data = df2,
+        mapping = ggplot2::aes(
+          x = condition,
+          y = 1.05,
+          label = significance,
+          fill = NULL
+        ),
+        size = 5,
+        na.rm = TRUE
+      )
+  }
 
-    # adding significance labels to pie charts for grouped proportion tests
-    if (isTRUE(facet.proptest)) {
-      p <-
-        p +
-        ggplot2::geom_text(data = df2,
-                aes(x = condition, y = -0.10, label = significance, fill = NULL),
-                size = 3,
-                fontface = "bold"
-                )
-    }
-
-    # adding sample size info
-    if (isTRUE(sample.size.label)) {
-      p <-
-        p +
-        ggplot2::geom_text(data = df_n_label,
-                aes(x = condition, y = -0.05, label = N, fill = NULL),
-                size = 3)
-    }
-
+  # adding sample size info
+  if (isTRUE(sample.size.label)) {
+    p <-
+      p +
+      ggplot2::geom_text(
+        data = df_n_label,
+        mapping = ggplot2::aes(
+          x = condition,
+          y = -0.05,
+          label = N,
+          fill = NULL
+        ),
+        size = 4,
+        na.rm = TRUE
+      )
+  }
 
   # =========================== putting all together ========================
-  # if we need to modify x axis orientation
-    if (!base::missing(x.axis.orientation)) {
-      if (x.axis.orientation == "slant") {
-        p <- p + ggplot2::theme(axis.text.x = ggplot2::element_text( angle = 45, vjust = 1, hjust= 1, face = "bold"))
-      } else if (x.axis.orientation == "vertical") {
-        p <- p + ggplot2::theme(axis.text.x = ggplot2::element_text( angle = 90, vjust = 0.5, hjust= 1, face = "bold"))
-      }
+  # if we need to modify `x`-axis orientation
+  if (!base::missing(x.axis.orientation)) {
+    if (x.axis.orientation == "slant") {
+      p <-
+        p + ggplot2::theme(
+          axis.text.x = ggplot2::element_text(
+            angle = 45,
+            vjust = 1,
+            hjust = 1,
+            face = "bold"
+          )
+        )
+    } else if (x.axis.orientation == "vertical") {
+      p <-
+        p + ggplot2::theme(
+          axis.text.x = ggplot2::element_text(
+            angle = 90,
+            vjust = 0.5,
+            hjust = 1,
+            face = "bold"
+          )
+        )
     }
+  }
 
   # preparing the plot
   p <-
     p +
     ggplot2::labs(
+      x = xlab,
+      y = ylab,
       subtitle = subtitle,
       title = title,
       caption = caption
     )
 
-    # ---------------- adding ggplot component ---------------------------------
+  # ---------------- adding ggplot component ---------------------------------
 
-    # if any additional modification needs to be made to the plot
-    # this is primarily useful for grouped_ variant of this function
-    p <- p + ggplot.component
-
+  # if any additional modification needs to be made to the plot
+  # this is primarily useful for grouped_ variant of this function
+  p <- p + ggplot.component
 
   # return the final plot
   return(p)