+++ R/grouped_ggpiestats.R
@@ -24,34 +24,34 @@
 #' @inherit ggpiestats return return
 #'
 #' @examples
-#' 
+#'
 #' # grouped one-sample proportion tests
 #' ggstatsplot::grouped_ggpiestats(
 #'   data = mtcars,
 #'   grouping.var = am,
 #'   main = cyl
 #' )
-#' 
+#'
 #' # without condition and with count data
 #' library(jmv)
-#' 
+#'
 #' ggstatsplot::grouped_ggpiestats(
 #'   data = as.data.frame(HairEyeColor),
 #'   main = Hair,
 #'   counts = Freq,
 #'   grouping.var = Sex
 #' )
-#' 
+#'
 #' # the following will take slightly more amount of time
 #' \dontrun{
 #' # for reproducibility
 #' set.seed(123)
-#' 
+#'
 #' # let's create a smaller dataframe
 #' diamonds_short <- ggplot2::diamonds %>%
 #'   dplyr::filter(.data = ., cut %in% c("Fair", "Very Good", "Ideal")) %>%
 #'   dplyr::sample_frac(tbl = ., size = 0.10)
-#' 
+#'
 #' # plot
 #' ggstatsplot::grouped_ggpiestats(
 #'   data = diamonds_short,
@@ -101,6 +101,7 @@ grouped_ggpiestats <- function(data,
                                package = "RColorBrewer",
                                palette = "Dark2",
                                direction = 1,
+                               ggplot.component = NULL,
                                messages = TRUE,
                                ...) {
 
@@ -285,6 +286,7 @@ grouped_ggpiestats <- function(data,
       package = package,
       palette = palette,
       direction = direction,
+      ggplot.component = ggplot.component,
       messages = messages
     )